Refactor diagnostics to use EditorDiagnostic interface
Introduces DiagnosticsProvider, EditorDiagnostic, and JdtDiagnosticsProvider for unified diagnostics handling. Refactors JavaCodeEditorPane to use EditorDiagnostic instead of ProblemDiagnostic, removing duplicate logic and improving modularity and maintainability.
